Bust ticket preview/image cache for newly-updated tickets
On some browsers, updating the ticket image does not refresh the preview.
The preview actually gets updated server-side, but the browser just uses
the cached version :< Let's update the preview/image URL whenever the
ticket gets updated.

This ideally should have a separate field just for image uploads, but
for small-scale events this is just fine.