Reimplement the ticket rendering mechanism
Instead of the previous "lightweight" system, where the Django codebase
itself would glue together the preview, we now use a dedicated system
run from the background tasks queue.

Ticket generation now invokes a specified "ticket renderer" which is a
container image that accepts data to render in a temporary directory,
then outputs a single PNG image there. This PNG image is then saved to
our media storage. How that container implements the said rendering
does not matter - we invoke what the user wants and off it goes.

By default, the container runs with extremely limited privileges and no
network access - the image must already be available on the system which
runs the task queue, as arbitrary image pulls are forbidden as well.

The newly-rendered tickets include the proper nickname, numbers, etc -
they are essentially final and ready to print. The old system is nuked.