Add an interstitial "thank you" page after reserving a ticket
This is a bit of an emergency change, since it masks a nasty issue with
the P24 payment backend. P24 occasionally becomes *SLOW*, to the point
where registering a new payment can time out a request. In the current
flow, we redirected to the payment screen right after registering a
ticket, thus triggering the payment registration. This would time out,
people would try to refresh the registration page, and end up with 5+
tickets on their accounts.

This interstitial prevents registering a payment right away. Even if it
times out, people don't end up with too many unused tickets and invalid
"your ticket has been cancelled" messages.