Export applications using the application template column order
Until now, when applications were exported as XLSX, the column order was
"whatever Postgres stored it as for the first time", so kinda random.
With this, we first discover the initial column order based on the app
type template, and then optionally backfill the columns if any unknown
fields are encountered.